Refactor(service): 예약 상태 변경 서비스 객체를 추상화
- 기존의 예약 상태 변경 서비스 클래스를 점주용으로 리팩토링
