+++ src/main/java/com/prgms/allen/dining/domain/reservation/service/OwnerReservationStatusUpdateService.java
@@ -1,27 +1,31 @@
-package com.prgms.allen.dining.domain.reservation;
+package com.prgms.allen.dining.domain.reservation.service;
 
 import java.text.MessageFormat;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.prgms.allen.dining.domain.member.entity.MemberType;
 import com.prgms.allen.dining.domain.reservation.dto.ReservationStatusUpdateReq;
 import com.prgms.allen.dining.domain.reservation.entity.Reservation;
 
 @Service
 @Transactional
-public class ReservationStatusUpdateService {
+@Qualifier("ownerReservationStatusUpdateService")
+public class OwnerReservationStatusUpdateService implements ReservationStatusUpdateService {
 
-	private static final Logger log = LoggerFactory.getLogger(ReservationStatusUpdateService.class);
+	private static final Logger log = LoggerFactory.getLogger(OwnerReservationStatusUpdateService.class);
 
 	private final ReservationService reservationService;
 
-	public ReservationStatusUpdateService(ReservationService reservationService) {
+	public OwnerReservationStatusUpdateService(ReservationService reservationService) {
 		this.reservationService = reservationService;
 	}
 
+	@Override
 	public void update(Long reservationId, Long ownerId, ReservationStatusUpdateReq updateReq) {
 		switch (updateReq.status()) {
 			case CONFIRMED -> confirm(reservationId, ownerId);
@@ -43,7 +47,7 @@ public class ReservationStatusUpdateService {
 
 	private void cancel(Long reservationId, Long ownerId) {
 		Reservation findReservation = reservationService.findById(reservationId);
-		findReservation.cancel(ownerId);
+		findReservation.cancel(MemberType.OWNER, ownerId);
 		log.info("Reservation {}'s status updated to {}", reservationId, findReservation.getStatus());
 	}
